added more rate increments. rate is now an integer between -5 and 10, not a double.  Double didn't make sense anyway as it wasn't a percent, but just a range.

+++ trunk/amis/src/AmisGuiMFC2/include/gui/self-voicing/directshow/dx_audio_player.h
@@ -221,8 +221,8 @@ private:
 
 #ifdef WITH_TPB_AUDIO_SPEEDUP
   public:
-	void set_rate(double rate);
-	double get_rate();
+	void set_rate(int rate);
+	int get_rate();
 
   private:
 	void set_rate_values(double crossFadeSpeed, int crossFadeWindowLength,
